requirements.txt
python-dotenv==0.17.1
requests>= 2.18.4
ujson==4.0.2
dhooks==1.1.4
